rename BottomNavBar and incorporated TopNavBar into NavigationShell logic
